+++ src/sequences/fakeSequences/fakeLastNamesSequence.ts
@@ -1,9 +1,9 @@
 import faker from "faker";
 import * as vscode from "vscode";
 import { getHumanizedLanguageName } from "../../helpers/utils";
-import { AFakerSequenceBase } from "../fakerSequenceBase";
+import { FakerSequenceBase } from "../sequenceBase";
 
-export class FakeLastNamesSequence extends AFakerSequenceBase {
+export class FakeLastNamesSequence extends FakerSequenceBase {
 	public get name(): string {
 		return vscode.l10n.t("{0} last names", getHumanizedLanguageName(this.locale));
 	}